+++ operator/src/main/java/org/keycloak/operator/v2alpha1/crds/realm/RealmStatus.java
@@ -16,18 +16,40 @@
  */
 package org.keycloak.operator.v2alpha1.crds.realm;
 
-import io.fabric8.kubernetes.api.model.Namespaced;
-import io.fabric8.kubernetes.client.CustomResource;
-import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Plural;
-import io.fabric8.kubernetes.model.annotation.ShortNames;
-import io.fabric8.kubernetes.model.annotation.Version;
-import org.keycloak.operator.Constants;
+public class RealmStatus {
+    public enum State {
+        DONE,
+        ERROR,
+        STARTED,
+        UNKNOWN
+    }
 
-@Group(Constants.CRDS_GROUP)
-@Version(Constants.CRDS_VERSION)
-@ShortNames(Constants.REALM_SHORT_NAME)
-@Plural(Constants.REALM_PLURAL_NAME)
-public class RealmImporter extends CustomResource<RealmImporterSpec, RealmImporterStatus> implements Namespaced {
+    private State state = State.UNKNOWN;
+    private boolean error;
+    private String message;
+
+    public State getState() {
+        return state;
+    }
+
+    public void setState(State state) {
+        this.state = state;
+    }
+
+    public boolean isError() {
+        return error;
+    }
+
+    public void setError(boolean error) {
+        this.error = error;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
 
 }
